pyproject.toml
[project]
name = "uc-data-advisor"
version = "0.1.0"
description = "Unity Catalog Data Advisor Agent for natural language dataset discovery"
requires-python = ">=3.12"
dependencies = [
    "databricks-sdk>=0.99.0",
    "mlflow>=3.10.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
